view acitvate courses
